requirements.txt
gradio>=4.0.0
Pillow>=9.0.0
aisuite>=0.1.0
python-dotenv>=0.19.0
docstring_parser>=0.16 